+++ crates/hyperswitch_domain_models/src/router_request_types.rs
@@ -67,7 +67,7 @@ pub struct PaymentsAuthorizeData {
 #[derive(Debug, serde::Deserialize, Clone)]
 pub struct PaymentCharges {
     pub charge_type: api_models::enums::PaymentChargeType,
-    pub fees: i64,
+    pub fees: MinorUnit,
     pub transfer_account_id: String,
 }
 
@@ -241,9 +241,13 @@ pub struct PaymentsPreProcessingData {
     pub surcharge_details: Option<SurchargeDetails>,
     pub browser_info: Option<BrowserInformation>,
     pub connector_transaction_id: Option<String>,
+    pub enrolled_for_3ds: bool,
     pub mandate_id: Option<api_models::payments::MandateIds>,
     pub related_transaction_id: Option<String>,
     pub redirect_response: Option<CompleteAuthorizeRedirectResponse>,
+
+    // New amount for amount frame work
+    pub minor_amount: Option<MinorUnit>,
 }
 
 impl TryFrom<PaymentsAuthorizeData> for PaymentsPreProcessingData {
@@ -253,6 +257,7 @@ impl TryFrom<PaymentsAuthorizeData> for PaymentsPreProcessingData {
         Ok(Self {
             payment_method_data: Some(data.payment_method_data),
             amount: Some(data.amount),
+            minor_amount: Some(data.minor_amount),
             email: data.email,
             currency: Some(data.currency),
             payment_method_type: data.payment_method_type,
@@ -268,6 +273,7 @@ impl TryFrom<PaymentsAuthorizeData> for PaymentsPreProcessingData {
             mandate_id: data.mandate_id,
             related_transaction_id: data.related_transaction_id,
             redirect_response: None,
+            enrolled_for_3ds: data.enrolled_for_3ds,
         })
     }
 }
@@ -279,6 +285,7 @@ impl TryFrom<CompleteAuthorizeData> for PaymentsPreProcessingData {
         Ok(Self {
             payment_method_data: data.payment_method_data,
             amount: Some(data.amount),
+            minor_amount: Some(data.minor_amount),
             email: data.email,
             currency: Some(data.currency),
             payment_method_type: None,
@@ -294,6 +301,7 @@ impl TryFrom<CompleteAuthorizeData> for PaymentsPreProcessingData {
             mandate_id: data.mandate_id,
             related_transaction_id: None,
             redirect_response: data.redirect_response,
+            enrolled_for_3ds: true,
         })
     }
 }
@@ -638,7 +646,7 @@ pub struct PayoutsData {
     pub connector_payout_id: Option<String>,
     pub destination_currency: storage_enums::Currency,
     pub source_currency: storage_enums::Currency,
-    pub payout_type: storage_enums::PayoutType,
+    pub payout_type: Option<storage_enums::PayoutType>,
     pub entity_type: storage_enums::PayoutEntityType,
     pub customer_details: Option<CustomerDetails>,
     pub vendor_details: Option<api_models::payouts::PayoutVendorAccountDetails>,