+++ crates/router/src/core/payments/flows.rs
@@ -988,7 +988,6 @@ default_imp_for_pre_processing_steps!(
     connector::Prophetpay,
     connector::Rapyd,
     connector::Riskified,
-    connector::Shift4,
     connector::Signifyd,
     connector::Square,
     connector::Stax,