+++ crates/router/src/core/payments/helpers.rs
@@ -2305,6 +2305,7 @@ pub fn validate_payment_method_type_against_payment_method(
                 | api_enums::PaymentMethodType::Eps
                 | api_enums::PaymentMethodType::BancontactCard
                 | api_enums::PaymentMethodType::Blik
+                | api_enums::PaymentMethodType::LocalBankRedirect
                 | api_enums::PaymentMethodType::OnlineBankingThailand
                 | api_enums::PaymentMethodType::OnlineBankingCzechRepublic
                 | api_enums::PaymentMethodType::OnlineBankingFinland
@@ -2348,6 +2349,13 @@ pub fn validate_payment_method_type_against_payment_method(
         api_enums::PaymentMethod::Reward => matches!(
             payment_method_type,
             api_enums::PaymentMethodType::Evoucher | api_enums::PaymentMethodType::ClassicReward
+        ),
+        api_enums::PaymentMethod::RealTimePayment => matches!(
+            payment_method_type,
+            api_enums::PaymentMethodType::Fps
+                | api_enums::PaymentMethodType::DuitNow
+                | api_enums::PaymentMethodType::PromptPay
+                | api_enums::PaymentMethodType::VietQr
         ),
         api_enums::PaymentMethod::Upi => matches!(
             payment_method_type,
@@ -3343,6 +3351,7 @@ pub fn router_data_type_conversion<F1, F2, Req1, Req2, Res1, Res2>(
         merchant_id: router_data.merchant_id,
         address: router_data.address,
         amount_captured: router_data.amount_captured,
+        minor_amount_captured: router_data.minor_amount_captured,
         auth_type: router_data.auth_type,
         connector: router_data.connector,
         connector_auth_type: router_data.connector_auth_type,
@@ -3879,6 +3888,9 @@ pub async fn get_additional_payment_data(
         api_models::payments::PaymentMethodData::Reward => {
             api_models::payments::AdditionalPaymentData::Reward {}
         }
+        api_models::payments::PaymentMethodData::RealTimePayment(_) => {
+            api_models::payments::AdditionalPaymentData::RealTimePayment {}
+        }
         api_models::payments::PaymentMethodData::Upi(_) => {
             api_models::payments::AdditionalPaymentData::Upi {}
         }
@@ -4015,12 +4027,13 @@ pub fn get_applepay_metadata(
         })
 }
 
+#[cfg(all(feature = "retry", feature = "connector_choice_mca_id"))]
 pub async fn get_apple_pay_retryable_connectors<F>(
     state: SessionState,
     merchant_account: &domain::MerchantAccount,
     payment_data: &mut PaymentData<F>,
     key_store: &domain::MerchantKeyStore,
-    decided_connector_data: api::ConnectorData,
+    pre_routing_connector_data_list: &[api::ConnectorData],
     merchant_connector_id: Option<&String>,
 ) -> CustomResult<Option<Vec<api::ConnectorData>>, errors::ApiErrorResponse>
 where
@@ -4035,13 +4048,17 @@ where
             field_name: "profile_id",
         })?;
 
+    let pre_decided_connector_data_first = pre_routing_connector_data_list
+        .first()
+        .ok_or(errors::ApiErrorResponse::IncorrectPaymentMethodConfiguration)?;
+
     let merchant_connector_account_type = get_merchant_connector_account(
         &state,
         merchant_account.merchant_id.as_str(),
         payment_data.creds_identifier.to_owned(),
         key_store,
-        profile_id, // need to fix this
-        &decided_connector_data.connector_name.to_string(),
+        profile_id,
+        &pre_decided_connector_data_first.connector_name.to_string(),
         merchant_connector_id,
     )
     .await?;
@@ -4063,9 +4080,14 @@ where
             .await
             .to_not_found_response(errors::ApiErrorResponse::InternalServerError)?;
 
-        let mut connector_data_list = vec![decided_connector_data.clone()];
+        let profile_specific_merchant_connector_account_list = filter_mca_based_on_business_profile(
+            merchant_connector_account_list,
+            Some(profile_id.to_string()),
+        );
 
-        for merchant_connector_account in merchant_connector_account_list {
+        let mut connector_data_list = vec![pre_decided_connector_data_first.clone()];
+
+        for merchant_connector_account in profile_specific_merchant_connector_account_list {
             if is_apple_pay_simplified_flow(
                 merchant_connector_account.metadata,
                 merchant_connector_account
@@ -4090,7 +4112,48 @@ where
                 }
             }
         }
-        Some(connector_data_list)
+
+        let fallback_connetors_list = crate::core::routing::helpers::get_merchant_default_config(
+            &*state.clone().store,
+            profile_id,
+            &api_enums::TransactionType::Payment,
+        )
+        .await
+        .change_context(errors::ApiErrorResponse::InternalServerError)
+        .attach_printable("Failed to get merchant default fallback connectors config")?;
+
+        let mut routing_connector_data_list = Vec::new();
+
+        pre_routing_connector_data_list.iter().for_each(|pre_val| {
+            routing_connector_data_list.push(pre_val.merchant_connector_id.clone())
+        });
+
+        fallback_connetors_list.iter().for_each(|fallback_val| {
+            routing_connector_data_list
+                .iter()
+                .all(|val| *val != fallback_val.merchant_connector_id)
+                .then(|| {
+                    routing_connector_data_list.push(fallback_val.merchant_connector_id.clone())
+                });
+        });
+
+        // connector_data_list is the list of connectors for which Apple Pay simplified flow is configured.
+        // This list is arranged in the same order as the merchant's connectors routingconfiguration.
+
+        let mut ordered_connector_data_list = Vec::new();
+
+        routing_connector_data_list
+            .iter()
+            .for_each(|merchant_connector_id| {
+                let connector_data = connector_data_list.iter().find(|connector_data| {
+                    *merchant_connector_id == connector_data.merchant_connector_id
+                });
+                if let Some(connector_data_details) = connector_data {
+                    ordered_connector_data_list.push(connector_data_details.clone());
+                }
+            });
+
+        Some(ordered_connector_data_list)
     } else {
         None
     };
@@ -4314,6 +4377,11 @@ pub fn get_key_params_for_surcharge_details(
         )),
         api_models::payments::PaymentMethodData::MandatePayment => None,
         api_models::payments::PaymentMethodData::Reward => None,
+        api_models::payments::PaymentMethodData::RealTimePayment(real_time_payment) => Some((
+            common_enums::PaymentMethod::RealTimePayment,
+            real_time_payment.get_payment_method_type(),
+            None,
+        )),
         api_models::payments::PaymentMethodData::Upi(upi_data) => Some((
             common_enums::PaymentMethod::Upi,
             upi_data.get_payment_method_type(),