+++ crates/router/src/types.rs
@@ -809,6 +809,7 @@ impl<F1, F2, T1, T2> ForeignFrom<(&RouterData<F1, T1, PaymentsResponseData>, T2)
             connector_meta_data: data.connector_meta_data.clone(),
             connector_wallets_details: data.connector_wallets_details.clone(),
             amount_captured: data.amount_captured,
+            minor_amount_captured: data.minor_amount_captured,
             access_token: data.access_token.clone(),
             response: data.response.clone(),
             payment_id: data.payment_id.clone(),
@@ -842,13 +843,13 @@ impl<F1, F2, T1, T2> ForeignFrom<(&RouterData<F1, T1, PaymentsResponseData>, T2)
 #[cfg(feature = "payouts")]
 impl<F1, F2>
     ForeignFrom<(
-        &&mut RouterData<F1, PayoutsData, PayoutsResponseData>,
+        &RouterData<F1, PayoutsData, PayoutsResponseData>,
         PayoutsData,
     )> for RouterData<F2, PayoutsData, PayoutsResponseData>
 {
     fn foreign_from(
         item: (
-            &&mut RouterData<F1, PayoutsData, PayoutsResponseData>,
+            &RouterData<F1, PayoutsData, PayoutsResponseData>,
             PayoutsData,
         ),
     ) -> Self {
@@ -870,6 +871,7 @@ impl<F1, F2>
             connector_meta_data: data.connector_meta_data.clone(),
             connector_wallets_details: data.connector_wallets_details.clone(),
             amount_captured: data.amount_captured,
+            minor_amount_captured: data.minor_amount_captured,
             access_token: data.access_token.clone(),
             response: data.response.clone(),
             payment_id: data.payment_id.clone(),