+++ cypress-tests/cypress/e2e/PaymentTest/00016-BankTransfers.cy.js
@@ -6,6 +6,8 @@ import getConnectorDetails, * as utils from "../PaymentUtils/utils";
 let globalState;
 
 describe("Bank Transfers", () => {
+  let should_continue = true; // variable that will be used to skip tests if a previous test fails
+
   before("seed global state", () => {
     cy.task("getGlobalState").then((state) => {
       globalState = new State(state);
@@ -16,15 +18,13 @@ describe("Bank Transfers", () => {
     cy.task("setGlobalState", globalState.data);
   });
 
+  beforeEach(function () {
+    if (!should_continue) {
+      this.skip();
+    }
+  });
+
   context("Bank transfer - Pix forward flow", () => {
-    let should_continue = true; // variable that will be used to skip tests if a previous test fails
-
-    beforeEach(function () {
-      if (!should_continue) {
-        this.skip();
-      }
-    });
-
     it("create-payment-call-test", () => {
       let data = getConnectorDetails(globalState.get("connectorId"))[
         "bank_transfer_pm"
@@ -37,7 +37,7 @@ describe("Bank Transfers", () => {
         res_data,
         "three_ds",
         "automatic",
-        globalState,
+        globalState
       );
       if (should_continue)
         should_continue = utils.should_continue_further(res_data);
@@ -58,7 +58,7 @@ describe("Bank Transfers", () => {
         req_data,
         res_data,
         true,
-        globalState,
+        globalState
       );
       if (should_continue)
         should_continue = utils.should_continue_further(res_data);
@@ -70,7 +70,7 @@ describe("Bank Transfers", () => {
       cy.handleBankTransferRedirection(
         globalState,
         payment_method_type,
-        expected_redirection,
+        expected_redirection
       );
     });
   });